ci: notify_maintainers: simplify scripts and use actions/github-script
Attempting once more to fix a 403 error in the notify_maintainers job.
Simplify the notify_maintainers.py script which now assumes the GitHub
environment with parameters passed as environment variables only and
make it output the message to stdout. It is then the responsibility of
the notify_maintainers job to post it, via the actions/github-script
action. In the tests I performed, the comment was successfully posted
by "github-actions (bot)".

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>