ci: code_style: run directly in the Ubuntu VM (no container)
The code_style job doesn't have so many special dependencies so it can
as well be run directly in the Ubuntu VM provided by the GitHub Actions
environment. With this change, the job typically takes less than 30
seconds to complete instead of 1 minute. Not a massive impact on the
overall CI workflow, but still good to take if only for the "keep it
simple" principle.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>