ci: free disk space earlier and in all jobs
Address two new "no space left on device" errors in CI:

- The first one appeared in the QEMUv8_checks_image_build job. To fix
it, avoid cloning the OP-TEE repository which is not needed in that job
i.e., remove the checkout action, and also run the host cleanup script.
The two other "image_build" jobs are updated as well for consistency.

- The second one happens in the _checks jobs. The fix consists in
running the host cleanup script earlier, before loading the Docker
image.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>