core: tee_ree_fs: conditionally remove corrupt dirf.db file
During the creation of the OP-TEE REE-FS database file, several
RPC commands are executed. If an unexpected power outage occurs
during the initial database setup in mass production, an incomplete
dirf.db file with a size of 0 bytes may be left behind.

This patch updates the flow to automatically remove an empty
dirf.db file if no hash value is set, before returning an error
code. This allows the caller to retry the secure store operation,
which will then recreate dirf.db as needed.

Link: https://github.com/OP-TEE/optee_os/issues/7512
Signed-off-by: Ox Yeh <ox.yeh@mediatek.com>