core: tee_ree_fs: remove corrupt file without rollback protection
During the creation of the OP-TEE REE-FS database file, several
RPC commands are executed. If an unexpected power outage occurs
during this process, it may result in an incomplete dirf.db file
with a size of 0 bytes, and this file has not yet been configured
with rollback protection.

This change extends the error handling in ree_fs_open_primitive
function to conditionally remove the corrupted file when rollback
protection is not set, allowing the caller to recreate the file
later. This also resolves the previously mentioned dirf.db issue.

Link: https://github.com/OP-TEE/optee_os/issues/7512
Signed-off-by: Ox Yeh <ox.yeh@mediatek.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@st.com>