ci: notify_maintainers.py: filter out comment authors
This is a small bug fix for the notify_maintainers.py script which
should not mention a user who has already been involved in the
discussion. Currently, it filters out users that have explicitly been
mentioned (via '@user'), either by the CI bot or by a real user. But
it does not consider the comment authors. This is a mistake. Update
the code accordingly.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>